Make keystore properties cast nullable
In groovy there is no need to cast to string but in kotlin you cast keystoreProperties["XXXX"] to String which fail on null, there is two solution force every user who want to contribute to have a keystore or allow nullable keystoreProperties.
I went with the latter to limit contributor friction.